:sparkles: feat (new_claim): Add API endpoint for creating new klaim
This code change adds a new API endpoint in the NewKlaimController to handle requests for creating a new klaim using the E-KLAIM API. The request data is validated to ensure all required information is present and in the correct format. Once validated, the data is packaged in JSON format and sent to the E-KLAIM API using the EklaimService. This change improves the functionality of the application by allowing users to create new klaims through the API.